Subject: 1st draft of the code for text generation based on 1 word
this is the1st draft I worked on to be able to display a serie of
words based on a text used for model training. There are several
functions that I developed that will have to be improved and I trained
only 1 epoch with no steps to make sure it works 1st. Without GPU this
is too long